+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     * Fixed error message showing when loading a dotLottie without custom metadata
     * Fixed loading image assets from dotLottie
     * Saving a file sets the default export path
+    * Importing non-animated SVG uses the same duration as the current animation
 * Scripting
     * `window.choose_option()` method that shows a dilaog to display a selection
     * Exception from I/O plugins now properly show in the script console